QDateTime: Store Julian Day as qint64
Store the QDate Julian Day number as an qint64 instead of uint32 to
enable support for dates before 2 January 4713 BCE.  This changes the
possible date range to be approx 2.5 Quadrillion BC to 2.5 Quadrillion
AD. A qint32 was not used as it only covers 5 million BCE to 5 million
CE which does include Geological or Astronomical time.

The effective supported date range is currently 4800 BCE to 1.4 million
CE due to restrictions in existing conversion formulas.  The effective
range will be extended later with new formulas.

+++ tests/auto/corelib/tools/qdatetime/tst_qdatetime.cpp
@@ -883,22 +883,22 @@ void tst_QDateTime::daysTo()
     QDateTime dt2(QDate(1760, 2, 2), QTime());
     QDateTime dt3(QDate(1760, 3, 2), QTime());
 
-    QCOMPARE(dt1.daysTo(dt2), 31);
+    QCOMPARE(dt1.daysTo(dt2), (qint64) 31);
     QCOMPARE(dt1.addDays(31), dt2);
 
-    QCOMPARE(dt2.daysTo(dt3), 29);
+    QCOMPARE(dt2.daysTo(dt3), (qint64) 29);
     QCOMPARE(dt2.addDays(29), dt3);
 
-    QCOMPARE(dt1.daysTo(dt3), 60);
+    QCOMPARE(dt1.daysTo(dt3), (qint64) 60);
     QCOMPARE(dt1.addDays(60), dt3);
 
-    QCOMPARE(dt2.daysTo(dt1), -31);
+    QCOMPARE(dt2.daysTo(dt1), (qint64) -31);
     QCOMPARE(dt2.addDays(-31), dt1);
 
-    QCOMPARE(dt3.daysTo(dt2), -29);
+    QCOMPARE(dt3.daysTo(dt2), (qint64) -29);
     QCOMPARE(dt3.addDays(-29), dt2);
 
-    QCOMPARE(dt3.daysTo(dt1), -60);
+    QCOMPARE(dt3.daysTo(dt1), (qint64) -60);
     QCOMPARE(dt3.addDays(-60), dt1);
 }
